Extracts color fill logic into a reusable helper function
Refactors color fill determination logic into a dedicated helper function to improve code reuse and maintainability. Updates references in test files to utilize the new helper.

Promotes cleaner separation of concerns and reduces duplication.
